refactor: saveResourcePage api function
this commit first moves the `saveResourcePage` function into the
api file, and then creates supporting api functions for creating,
updating, and deleting resource pages
