+++ src/api.js
@@ -1,5 +1,11 @@
 import axios from 'axios';
-import { getNavFolderDropdownFromFolderOrder, parseDirectoryFile } from './utils';
+import { Base64 } from 'js-base64';
+import {
+    getNavFolderDropdownFromFolderOrder,
+    parseDirectoryFile,
+    generateResourceFileName,
+    concatFrontMatterMdBody,
+} from './utils';
 
 // axios settings
 axios.defaults.withCredentials = true
@@ -140,6 +146,95 @@ const moveFiles = async (siteName, selectedFiles, title, parentFolder) => {
     return await axios.post(`${baseApiUrl}/move/${newPath}`, params)
 }
 
+// Resource pages
+const createNewResourcePage = async ({ content, pageName, siteName, category }) => {
+    const params = {
+        content,
+        pageName,
+    };
+    await axios.post(`${process.env.REACT_APP_BACKEND_URL}/sites/${siteName}/resources/${category}/pages`, params);
+}
+
+const deleteResourcePage = async ({ sha, siteName, category, fileName }) => {
+    const params = {
+        data: {
+          sha,
+        },
+    }
+    await axios.delete(`${process.env.REACT_APP_BACKEND_URL}/sites/${siteName}/resources/${category}/pages/${fileName}`, params);
+}
+
+const updateResourcePage = async ({ sha, content, siteName, category, fileName }) => {
+    const params = {
+        sha,
+        content,
+    }
+    await axios.post(`${process.env.REACT_APP_BACKEND_URL}/sites/${siteName}/resources/${category}/pages/${fileName}`, params);
+}
+
+const saveResourcePage = async (fileInfo) => {
+    const {
+      category,
+      date,
+      fileUrl,
+      mdBody,
+      permalink,
+      resourceType,
+      sha,
+      title,
+      fileName,
+      isNewFile,
+      siteName,
+    } = fileInfo
+
+    const newFileName = generateResourceFileName(title.toLowerCase(), date, resourceType);
+    const frontMatter = { title, date };
+
+    if (permalink) {
+      frontMatter.permalink = permalink;
+    }
+
+    if (fileUrl) {
+      frontMatter.file_url = fileUrl;
+    }
+
+    const content = concatFrontMatterMdBody(frontMatter, mdBody);
+    const base64EncodedContent = Base64.encode(content);
+
+    if (newFileName !== fileName) {
+      // We'll need to create a new .md file with a new filename
+      await createNewResourcePage({
+        content: base64EncodedContent,
+        pageName: newFileName,
+        siteName,
+        category,
+      });
+
+      // If it is an existing file, delete the existing page
+      if (!isNewFile) {
+        await deleteResourcePage({ sha, siteName, category, fileName})
+      }
+    } else {
+      // Save to existing .md file
+      await updateResourcePage({
+        content: base64EncodedContent,
+        sha,
+        siteName,
+        category,
+        fileName,
+      })
+    }
+
+    let newPageUrl
+    if (resourceType === 'file') {
+      newPageUrl = `/sites/${siteName}/resources/${category}`
+    } else {
+      newPageUrl = `/sites/${siteName}/resources/${category}/${newFileName}`
+    }
+
+    return newPageUrl
+}
+
 export {
     getDirectoryFile,
     setDirectoryFile,
@@ -155,4 +250,5 @@ export {
     updatePage,
     deletePage,
     moveFiles,
+    saveResourcePage,
 }
